moved tests to their own folder to avoid creating declaration files
