+++ .eslintrc.js
@@ -12,7 +12,7 @@ module.exports = {
   ],
   parser: "@typescript-eslint/parser",
   parserOptions: {
-    project: "./tsconfig.base.json",
+    project: "./tsconfig.json",
     sourceType: "module",
   },
   plugins: ["@typescript-eslint", "prettier"],
