+++ tsconfig.json
@@ -1,4 +1,15 @@
 {
-  "extends": "./tsconfig.base.json",
-  "exclude": ["./src/**/*.test.*ts"]
+  "compilerOptions": {
+    "module": "esnext",
+    "moduleResolution": "node",
+    "declaration": true,
+    "declarationDir": "./types",
+    "noImplicitAny": true,
+    "strictNullChecks": true,
+    "removeComments": true,
+    "sourceMap": true,
+    "allowJs": false
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["./node_modules/*", "./dist/*"]
 }
